requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
colorama==0.4.4
fastapi==0.68.2
h11==0.12.0
httptools==0.2.0
idna==3.2
pydantic==1.8.2
python-dotenv==0.19.0
PyYAML==5.4.1
requests==2.26.0
starlette==0.14.2
typing-extensions==3.10.0.2
urllib3==1.26.7
uvicorn==0.15.0
watchgod==0.7
websockets==10.0
